refator: 리덕스 초기값 세팅 Main에서 할당하도록 수정
